+++ core/shared/src/main/scala/cats/effect/package.scala
@@ -76,4 +76,7 @@ package object effect {
 
   type Ref[F[_], A] = cekernel.Ref[F, A]
   val Ref = cekernel.Ref
+
+  type Semaphore[F[_]] = cekernel.Semaphore[F]
+  val Semaphore = cekernel.Semaphore
 }
